backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx>=0.27.0
sse-starlette>=2.1.3
google-generativeai>=0.8.0
chromadb>=0.5.5
reportlab>=4.2.2
python-multipart>=0.0.9
tenacity>=9.0.0
orjson>=3.10.7
pytest>=8.3.0
